Increase default errorlog bugger size
If logging is initialised (for example by using softIocPVX) before you have
a chance to set the error log buffer size, it is possible to end up with
messages that are truncated too early. This sets the default buffer and message
sizes larger in order to better handle this.

Note that this preserves the roughtly 5:1 ratio of buffer:message sizes, which
is necessary to ensure that messages are  not lost.